+++ RevenueCat/Plugins/iOS/PurchasesUnityHelper.m
@@ -59,17 +59,23 @@ char *makeStringCopy(NSString *nstring) {
             gameObject:(NSString *)gameObject
           observerMode:(BOOL)observerMode
  userDefaultsSuiteName:(nullable NSString *)userDefaultsSuiteName
-             useAmazon:(BOOL)useAmazon {
-    [[RCPurchases sharedPurchases] setDelegate:nil];
+             useAmazon:(BOOL)useAmazon
+     dangerousSettings:(NSDictionary *)dangerousSettingsAsDictionary {
+    if (RCPurchases.isConfigured) {
+        [[RCPurchases sharedPurchases] setDelegate:nil];
+    }
     self.products = nil;
     self.gameObject = nil;
+
+    RCDangerousSettings *dangerousSettings = [self dangerousSettingsFromDictionary:dangerousSettingsAsDictionary];
 
     [RCPurchases configureWithAPIKey:apiKey
                            appUserID:appUserID
                         observerMode:observerMode
                userDefaultsSuiteName:userDefaultsSuiteName
                       platformFlavor:self.platformFlavor
-               platformFlavorVersion:self.platformFlavorVersion];
+               platformFlavorVersion:self.platformFlavorVersion
+                   dangerousSettings:dangerousSettings];
     
     self.gameObject = gameObject;
     [[RCPurchases sharedPurchases] setDelegate:self];
@@ -180,6 +186,12 @@ char *makeStringCopy(NSString *nstring) {
         
         [self sendJSONObject:response toMethod:GET_OFFERINGS];
     }];
+}
+
+- (void)syncObserverModeAmazonPurchase:(NSString *)productID 
+                             receiptID:(NSString *)receiptID
+                          amazonUserID:(NSString *)amazonUserID {
+    // noop
 }
 
 - (void)setDebugLogsEnabled:(BOOL)enabled {
@@ -361,6 +373,15 @@ char *makeStringCopy(NSString *nstring) {
     return @"3.2.0-amazon.alpha";
 }
 
+- (nullable RCDangerousSettings *)dangerousSettingsFromDictionary:(NSDictionary *)dangerousSettingsAsDictionary {
+    RCDangerousSettings *dangerousSettings = nil;
+    if (dangerousSettingsAsDictionary.count != 0) {
+        BOOL autoSyncPurchases = [dangerousSettingsAsDictionary[@"AutoSyncPurchases"] boolValue];
+        dangerousSettings = [[RCDangerousSettings alloc]initWithAutoSyncPurchases:autoSyncPurchases];
+    }
+    return dangerousSettings;
+}
+
 @end
 
 #pragma mark Bridging Methods
@@ -378,12 +399,26 @@ void _RCSetupPurchases(const char *gameObject,
                        const char *apiKey,
                        const char *appUserID,
                        const BOOL observerMode,
-                       const char *userDefaultsSuiteName) {
+                       const char *userDefaultsSuiteName,
+                       const BOOL useAmazon,
+                       const char *dangerousSettingsJSON) {
+    NSError *error = nil;
+    NSData *dangerousSettingsData = [convertCString(dangerousSettingsJSON) dataUsingEncoding:NSUTF8StringEncoding];
+    NSDictionary *dangerousSettings = [NSJSONSerialization JSONObjectWithData:dangerousSettingsData
+                                                                      options:0
+                                                                        error:&error];
+    if (error) {
+        NSLog(@"Error parsing dangerousSettings JSON: %s %@", dangerousSettingsJSON, error.localizedDescription);
+        dangerousSettings = @{};
+    }
+
     [_RCUnityHelperShared() setupPurchases:convertCString(apiKey)
                                  appUserID:convertCString(appUserID)
                                 gameObject:convertCString(gameObject)
                               observerMode:observerMode
-                     userDefaultsSuiteName:convertCString(userDefaultsSuiteName)];
+                     userDefaultsSuiteName:convertCString(userDefaultsSuiteName)
+                                 useAmazon:useAmazon
+                         dangerousSettings:dangerousSettings];
 }
 
 void _RCGetProducts(const char *productIdentifiersJSON, const char *type) {
@@ -585,3 +620,8 @@ void _RCCanMakePayments(const char *featuresJSON) {
     [_RCUnityHelperShared() canMakePaymentsWithFeatures:canMakePaymentsRequest[@"features"]];
 }
 
+void _RCSyncObserverModeAmazonPurchase(const char *productID, const char *receiptID, const char *amazonUserID) {
+    [_RCUnityHelperShared() syncObserverModeAmazonPurchase:convertCString(productID)
+                                                 receiptID:convertCString(receiptID)
+                                              amazonUserID:convertCString(amazonUserID)];
+}