+++ RevenueCat/Scripts/PurchasesWrapperiOS.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 public class PurchasesWrapperiOS : IPurchasesWrapper
 {
     [DllImport("__Internal")]
-    private static extern void _RCSetupPurchases(string gameObject, string apiKey, string appUserId, bool observerMode, string userDefaultsSuiteName, bool useAmazon);
-    public void Setup(string gameObject, string apiKey, string appUserId, bool observerMode, string userDefaultsSuiteName, bool useAmazon)
+    private static extern void _RCSetupPurchases(string gameObject, string apiKey, string appUserId, bool observerMode, string userDefaultsSuiteName, bool useAmazon, string dangerousSettingsJson);
+    public void Setup(string gameObject, string apiKey, string appUserId, bool observerMode, string userDefaultsSuiteName, bool useAmazon, string dangerousSettingsJson)
     {
-        _RCSetupPurchases(gameObject, apiKey, appUserId, observerMode, userDefaultsSuiteName, useAmazon);
+        _RCSetupPurchases(gameObject, apiKey, appUserId, observerMode, userDefaultsSuiteName, useAmazon, dangerousSettingsJson);
     }
 
     [SuppressMessage("ReSharper", "NotAccessedField.Local")]
@@ -133,6 +133,13 @@ public class PurchasesWrapperiOS : IPurchasesWrapper
     public void GetOfferings()
     {
         _RCGetOfferings();
+    }
+
+    [DllImport("__Internal")]
+    private static extern void _RCSyncObserverModeAmazonPurchase(string productID, string receiptID, string amazonUserID);
+    public void SyncObserverModeAmazonPurchase(string productID, string receiptID, string amazonUserID)
+    {
+        _RCSyncObserverModeAmazonPurchase(productID, receiptID, amazonUserID);
     }
 
     [DllImport("__Internal")]